+++ tindahan_natin/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Sidebar for tablet/desktop devices
 - Remove `#` from the public store web links
+- Add product details button for public view
 
 ## [1.0.0+17] - 2026-06-02